requirements.txt
async-timeout==4.0.3
billiard==3.6.4.0
certifi==2023.11.17
charset-normalizer==2.0.12
coloredlogs==15.0.1
cycler==0.12.1
Deprecated==1.2.14
flatbuffers==23.5.26
fonttools==4.47.2
humanfriendly==10.0
idna==3.3
imageio==2.16.1
joblib==1.1.0
kiwisolver==1.4.0
kombu==5.3.5
matplotlib==3.5.1
mpmath==1.3.0
networkx==2.7.1
numpy==1.22.3
onnx==1.11.0
onnxruntime-gpu==1.14.1
opencv-python==4.8.0.76
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.0.1
prettytable==3.2.0
prompt-toolkit==3.0.43
protobuf==3.19.4
pydantic==1.10.14
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2023.4
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2022.2.9
tqdm==4.64.0
typing_extensions==4.4.0
urllib3==1.26.8
vine==5.1.0
wcwidth==0.2.5
wrapt==1.16.0